detail_pane: split into wrap, layout and entity siblings
Pure move. 1158 code lines to 702 in mod.rs; wrapping, column widths and
per-entity rows each get their own file.
