+++ mandible-tui/src/render/detail_pane/mod.rs
@@ -10,6 +10,13 @@
 //! hang-indent at the description column (spec §9.1a) and search must
 //! scroll to an exact line index (spec §10). `Wrap` stays on only as a
 //! defensive fallback against unclipped text (spec §9).
+mod entity;
+mod layout;
+mod wrap;
+
+use entity::*;
+use layout::*;
+use wrap::*;
 
 use crate::app::{App, Focus};
 use crate::glyphs::Glyphs;
@@ -278,28 +285,6 @@ fn render_verbatim(
     draw_clip_marker_rails(frame, inner, app.color_enabled, &clip_rows, scroll);
 }
 
-/// The visible window of `s` for horizontal scrolling of preformatted
-/// content: `offset` display-width columns trimmed off the left, capped to
-/// `width` columns of what remains (else it falls through to `Paragraph`'s
-/// defensive `Wrap` and gets reflowed, spec §9). Indexes character-by-
-/// character, never a byte offset (AGENTS.md); a double-width character
-/// straddling a boundary is dropped whole. Regression fixture: `ip`
-/// rendered through a real pty (AGENTS.md §3.2). Returns the window plus
-/// per-edge clip flags, fed to [`draw_clip_marker_rails`].
-fn hscroll_line(s: &str, offset: usize, width: usize) -> (Line<'static>, bool, bool) {
-    let total = display_width(s);
-    if offset == 0 && total <= width {
-        return (Line::from(s.to_string()), false, false);
-    }
-    if total <= offset || width == 0 {
-        return (Line::default(), false, false);
-    }
-    let clipped_left = offset > 0;
-    let clipped_right = total > offset + width;
-    let line = Line::from(hscroll_window(s, offset, width).into_owned());
-    (line, clipped_left, clipped_right)
-}
-
 /// Draw the per-line clip markers into the padding gutters either side of
 /// `inner` — the one blank column `Block::padding` leaves between border
 /// and content. `clip_rows` holds `(line index, clipped_left,
@@ -335,107 +320,6 @@ fn draw_clip_marker_rails(
             buf[(right_x, y)].set_symbol(">").set_style(marker);
         }
     }
-}
-
-/// Wrap one preformatted line to `width` display columns, losing nothing
-/// (wrap-mode counterpart of [`hscroll_line`], used with `[ui]
-/// horizontal_scroll` off, spec §9). Not [`wrap_words`]: that collapses
-/// interior spacing, destroying a tool's own column-aligned table (spec
-/// §4.1). A fitting line returns byte-identical; an over-wide one cuts at
-/// a whitespace boundary when one exists, else hard-cuts between
-/// characters; continuation rows carry the line's own indent, dropped
-/// past half the pane. Character-by-character, never a byte slice
-/// (AGENTS.md); [`width_prefix_end`] only returns real `char` boundaries.
-fn wrap_preformatted(line: &str, width: usize) -> Vec<String> {
-    let width = width.max(1);
-    if display_width(line) <= width {
-        return vec![line.to_string()];
-    }
-    let indent_width = display_width(&line[..line.len() - line.trim_start().len()]);
-    let hang = if indent_width * 2 < width {
-        " ".repeat(indent_width)
-    } else {
-        String::new()
-    };
-
-    let mut rows: Vec<String> = Vec::new();
-    let mut rest = line;
-    while !rest.is_empty() {
-        let prefix = if rows.is_empty() { "" } else { hang.as_str() };
-        let avail = width.saturating_sub(display_width(prefix)).max(1);
-        let mut cut = width_prefix_end(rest, avail);
-        if cut == 0 {
-            // A single character wider than the whole budget: take it
-            // anyway and overflow by the unavoidable minimum, exactly as
-            // `break_overlong_word` does, rather than loop forever.
-            cut = rest
-                .chars()
-                .next()
-                .map_or(rest.len(), |c: char| c.len_utf8());
-        }
-        if cut == rest.len() {
-            rows.push(format!("{prefix}{rest}"));
-            break;
-        }
-        // Prefer a whitespace boundary inside the window, so a word is not
-        // split when it did not have to be — but never one that would emit
-        // a row with no content on it.
-        let mut end = cut;
-        if let Some(pos) = rest[..cut].rfind(char::is_whitespace) {
-            if !rest[..pos].trim().is_empty() {
-                end = pos;
-            }
-        }
-        rows.push(format!("{prefix}{}", &rest[..end]));
-        rest = if end < cut {
-            // Broke at whitespace: that run was the break, not content.
-            rest[end..].trim_start()
-        } else {
-            &rest[end..]
-        };
-    }
-    rows
-}
-
-/// The byte index ending the longest prefix of `s` that fits `width`
-/// display columns — always a `char` boundary, so slicing at it cannot
-/// panic on multi-byte input.
-fn width_prefix_end(s: &str, width: usize) -> usize {
-    let mut used = 0usize;
-    for (idx, ch) in s.char_indices() {
-        let w = UnicodeWidthChar::width(ch).unwrap_or(0);
-        if used + w > width {
-            return idx;
-        }
-        used += w;
-    }
-    s.len()
-}
-
-fn hscroll_window(s: &str, offset: usize, width: usize) -> std::borrow::Cow<'_, str> {
-    if offset == 0 && display_width(s) <= width {
-        return std::borrow::Cow::Borrowed(s);
-    }
-    let mut remaining = offset;
-    let mut budget = width;
-    let mut result = String::new();
-    let mut trimming = offset > 0;
-    for ch in s.chars() {
-        let w = UnicodeWidthChar::width(ch).unwrap_or(0);
-        if trimming {
-            if w <= remaining {
-                remaining -= w;
-                continue;
-            }
-            trimming = false;
-        }
-        if w > budget {
-            break;
-        }
-        budget -= w;
-        result.push(ch);
-    }
-    std::borrow::Cow::Owned(result)
 }
 
 /// Recognize only the display-only prose prepended by `not_attested_fallback`.
@@ -839,77 +723,6 @@ fn looks_like_option_or_placeholder(word: &str) -> bool {
     has_letter && !word.chars().any(|c| c.is_lowercase())
 }
 
-/// Greedy word-wrap of `text` to at most `width` display columns per
-/// line, never breaking a word unless it alone exceeds `width`, in which
-/// case it breaks across as many lines as it takes ([`break_overlong_word`])
-/// rather than truncating: a truncated token is unrecoverable from the
-/// parsed view. Fixture: `smokecli unbreakable url`. Always returns at
-/// least one (possibly empty) chunk.
-fn wrap_words(text: &str, width: usize) -> Vec<String> {
-    let width = width.max(1);
-    let mut lines = Vec::new();
-    let mut current = String::new();
-    let mut current_width = 0usize;
-
-    for word in text.split_whitespace() {
-        let word_width = display_width(word);
-        let sep_width = usize::from(!current.is_empty());
-        if current_width + sep_width + word_width <= width {
-            if !current.is_empty() {
-                current.push(' ');
-            }
-            current.push_str(word);
-            current_width += sep_width + word_width;
-            continue;
-        }
-        if !current.is_empty() {
-            lines.push(std::mem::take(&mut current));
-            current_width = 0;
-        }
-        if word_width > width {
-            lines.extend(break_overlong_word(word, width));
-        } else {
-            current.push_str(word);
-            current_width = word_width;
-        }
-    }
-    if !current.is_empty() {
-        lines.push(current);
-    }
-    if lines.is_empty() {
-        lines.push(String::new());
-    }
-    lines
-}
-
-/// Break a single token wider than `width` display columns into as many
-/// width-limited chunks as it takes, so the token survives intact rather
-/// than being lost to an ellipsis truncation. Splits by summed
-/// [`unicode_width`], never byte index (AGENTS.md) and never `char` count
-/// (a double-width CJK/emoji character could overflow the line by one
-/// cell). A single character wider than `width` gets its own oversized
-/// chunk; no cut point inside a character exists.
-fn break_overlong_word(word: &str, width: usize) -> Vec<String> {
-    let width = width.max(1);
-    let mut chunks = Vec::new();
-    let mut current = String::new();
-    let mut current_width = 0usize;
-
-    for c in word.chars() {
-        let c_width = UnicodeWidthChar::width(c).unwrap_or(0);
-        if current_width + c_width > width && !current.is_empty() {
-            chunks.push(std::mem::take(&mut current));
-            current_width = 0;
-        }
-        current.push(c);
-        current_width += c_width;
-    }
-    if !current.is_empty() {
-        chunks.push(current);
-    }
-    chunks
-}
-
 /// A group heading with its source's terminator stripped: single-lined,
 /// trimmed, no trailing colon (`"GLOBAL OPTIONS:"`) or sentence-ending
 /// full stop (LVM's own per-stanza description used as a group label,
@@ -1007,152 +820,6 @@ fn group_divider_lead_line(
 ) -> Line<'static> {
     let label = truncate_to_width_marker(label, width, glyphs.ellipsis);
     Line::from(Span::styled(label, style::group_rule(palette)))
-}
-
-/// A spelling wider than this fraction of the pane does not get to set the
-/// shared column — it runs on past it instead, pushing its own first
-/// description line and nothing else (see [`SectionLayout`]). Mirrors the
-/// tree pane's summary-column rule (spec §9.1).
-const DESC_COLUMN_CAP_PERCENT: usize = 45;
-
-/// The share of a section's entities the shared column is fitted to
-/// (spec §9.3: "roughly the p90 spelling width — the majority, not the
-/// outliers").
-///
-/// Not the maximum. A column fitted to the widest spelling in the section
-/// is a column one entity chose for every other one, and the wider that
-/// entity is the less room the rest of the section's prose gets. Fitting
-/// the ninetieth percentile spends one extra line on the widest tenth and
-/// gives the width back to the other nine.
-const SHARED_COLUMN_PERCENTILE: usize = 90;
-
-/// The narrowest a description is allowed to be. A section's shared
-/// column is clamped down until this much of the pane is left for prose,
-/// however wide the section's heads are (spec §9.3). Measured against
-/// real output: at 20 columns `docker pull`'s `--platform` description
-/// breaks across six lines, one mid-word; at 28 it reads as prose.
-const MIN_DESC_WIDTH: usize = 28;
-
-/// Where a short spelling starts: the true left edge of the content area
-/// (spec §9.3). There is no uniform margin on a list section — the row's
-/// own shape decides which of the two columns it starts at.
-const SHORT_COLUMN: usize = 0;
-
-/// Where a long spelling starts, whether or not a short precedes it
-/// (spec §9.3): the display width of a short prefix, `-X, `.
-///
-/// A row that has a short renders it at [`SHORT_COLUMN`] and its long
-/// lands here by arithmetic; a row with no short is preindented to the
-/// same place. That is the whole point — the eye follows the longs down
-/// one column without having to know which rows happen to have a short
-/// letter as well.
-const LONG_COLUMN: usize = "-X, ".len();
-
-/// The indent POSITIONALS rows are inset by (spec §9.3): two columns, its
-/// own number and deliberately not [`LONG_COLUMN`], since coupling it to
-/// the flag columns would move two unrelated layouts together. MODIFIERS
-/// and ENVIRONMENT are bare-name sections too but stay laid out like
-/// FLAGS, against the content edge.
-const POSITIONAL_INDENT: usize = 2;
-
-/// The column an entity's spellings start at within a section indented by
-/// `indent` (spec §9.3). Shape decides it, never kind: a row whose first
-/// spelling is a short (or a dashless name) starts at the content edge; a
-/// long-only row is preindented to the same column a short row's long
-/// lands at. A row with more than two spellings (`-h, -?, -help, --help`)
-/// always flows from the short column, since there is no single "the
-/// long" in it to align to.
-fn spelling_column(entity: &Entity, indent: usize) -> usize {
-    indent + bare_spelling_column(entity)
-}
-
-/// [`spelling_column`] before the section's own indent is added.
-fn bare_spelling_column(entity: &Entity) -> usize {
-    if entity.spellings.len() > 2 {
-        return SHORT_COLUMN;
-    }
-    if entity
-        .spellings
-        .iter()
-        .any(|s| matches!(s.dashes, Dashes::None))
-    {
-        return SHORT_COLUMN;
-    }
-    if entity.short_spelling().is_some() {
-        return SHORT_COLUMN;
-    }
-    LONG_COLUMN
-}
-
-/// How a whole section is arranged. Chosen once per section, never per
-/// row, and per section rather than per pane (spec §9.3): positionals,
-/// flags, modifiers and environment variables have nothing to say to
-/// each other's widths. A row too wide for the column pushes only its
-/// own first description line right, one space past its head; every
-/// other line, and every other row, stays on the column.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct SectionLayout {
-    /// The section's shared description column: where every description
-    /// line in the section begins. The value placeholder is measured as
-    /// part of the spelling, not given its own aligned slot (spec §9.3).
-    description: usize,
-    /// Columns every row in the section is inset by — [`POSITIONAL_INDENT`]
-    /// for POSITIONALS, zero for the flag-shaped sections.
-    indent: usize,
-}
-
-/// The width that fits [`SHARED_COLUMN_PERCENTILE`] of `widths` — the
-/// smallest number at least that share of the entries are within.
-///
-/// Zero for an empty section, and the maximum for a section small enough
-/// that the percentile lands on it: a list of three flags aligns all three,
-/// because "the majority, not the outliers" only has anything to exclude
-/// once there is a tail to be an outlier in.
-fn percentile_width(widths: impl Iterator<Item = usize>) -> usize {
-    let mut widths: Vec<usize> = widths.collect();
-    if widths.is_empty() {
-        return 0;
-    }
-    widths.sort_unstable();
-    let n = widths.len();
-    let rank = (n * SHARED_COLUMN_PERCENTILE).div_ceil(100).max(1);
-    widths[rank - 1]
-}
-
-/// The layout for one section's `entities` in a pane `width` columns wide,
-/// inset by `indent`. The shared column is the lowest of three bounds:
-///
-/// 1. The percentile (spec §9.3): fitted to the majority, widest tenth
-///    excluded rather than clamped.
-/// 2. The pane cap (spec §9.1a): a head past [`DESC_COLUMN_CAP_PERCENT`]
-///    gets no vote.
-/// 3. The clamp (spec §9.3): comes down until [`MIN_DESC_WIDTH`] columns
-///    are left for prose.
-///
-/// ...floored at two past the deepest column a spelling can start at, so
-/// a description never lands left of the preindented longs it belongs to.
-fn section_layout(entities: &[&Entity], width: usize, indent: usize) -> SectionLayout {
-    let cap = width * DESC_COLUMN_CAP_PERCENT / 100;
-    let gap = 2;
-
-    // One measured width per row, from the pane's own left edge to the end
-    // of the row's placeholder: a preindented long is measured where it
-    // actually starts, and a placeholder is measured as part of the
-    // spelling it belongs to rather than against a slot of its own.
-    let fits = |w: usize| w + gap <= cap;
-    let fitting = entities
-        .iter()
-        .map(|e| entity_head_width(e, indent))
-        .filter(|w| fits(*w));
-
-    let floor = indent + LONG_COLUMN + gap;
-    let description = (percentile_width(fitting) + gap)
-        .min(width.saturating_sub(MIN_DESC_WIDTH))
-        .max(floor);
-    SectionLayout {
-        description,
-        indent,
-    }
 }
 
 /// One section's rendered body: its lines, its logical rows, and where a
@@ -1320,401 +987,6 @@ fn push_entity(
         lines: rendered.len(),
     });
     body.lines.extend(rendered);
-}
-
-/// An entity's spellings, e.g. `-i, --interactive` for a flag or
-/// `pathspec` for a positional — with a repeatable positional's `...`
-/// (spec §9.3).
-fn entity_name_spec(flag: &Entity) -> String {
-    // Every documented spelling, comma-separated, each reconstructed by
-    // `Spelling::render`: the dashes and the getopt_long `--[no-]foo`
-    // convention are display metadata, because the IR stores every name
-    // bare (`"help"`, never `"-help"`; `"foo"`, never `"[no-]foo"`) so
-    // that what a user searches and copies has no punctuation smuggled
-    // into it. This is the one place that spelling comes back together —
-    // and it is a *list*, so a row documenting four spellings
-    // (`-h, -?, -help, --help`) renders all four rather than the two a
-    // short/long pair could hold.
-    let spellings = flag
-        .spellings
-        .iter()
-        .map(Spelling::render)
-        .collect::<Vec<_>>()
-        .join(", ");
-    // `repeatable` is one fact spelled two ways by the two kinds'
-    // notation (`Entity::repeatable`): a flag says it by being accepted
-    // again (`-v -v -v`), a positional by the POSIX synopsis ellipsis the
-    // parser read it from (`FILE...`). Only the positional's notation is
-    // a suffix on the name, so only POSITIONALS gets one back — putting
-    // `...` on `-v` would invent a spelling nobody can type. Required/no
-    // marker is untouched; the ellipsis says "more than one", not
-    // "at least one".
-    if flag.kind == EntityKind::Positional && flag.repeatable {
-        return format!("{spellings}...");
-    }
-    spellings
-}
-
-/// How wide one entity's head runs: from the pane's left edge, through the
-/// section's indent and the column its shape starts it at, to the end of
-/// its value placeholder.
-///
-/// The single width the section's shared column is fitted to (spec §9.3).
-/// It has to be one number, because a placeholder measured separately is a
-/// placeholder the row is not charged for: `grep`'s `-e, --regexp
-/// PATTERNS` fits a column measured over `-e, --regexp` and overruns the
-/// one it is rendered against.
-fn entity_head_width(entity: &Entity, indent: usize) -> usize {
-    let mut width = spelling_column(entity, indent) + display_width(&entity_name_spec(entity));
-    if let Some(v) = entity_value_text(entity) {
-        let gap = if spelling_is_sigil(entity) { 0 } else { 1 };
-        width += gap + display_width(&v);
-    }
-    width
-}
-
-/// An entity's value placeholder, e.g. `FILE` or `[FILE]` when optional.
-/// `None` when it takes no value.
-fn entity_value_text(flag: &Entity) -> Option<String> {
-    flag.value_name
-        .as_ref()
-        .and_then(|name| match flag.value_kind {
-            ValueKind::Required => Some(name.clone()),
-            ValueKind::Optional => Some(format!("[{name}]")),
-            ValueKind::None => None,
-        })
-}
-
-/// True when this entity's value placeholder glues directly onto its
-/// spelling with no space — the argfile sigil flag's row-verbatim shape,
-/// `@<file>` (spec §4.5), rather than the ordinary `--output FILE` gap
-/// (spec §9.3). Decided by shape (a single dashless spelling whose first
-/// character is not alphanumeric), not by the literal `"@"`: a dashed
-/// short option like `-?` must not match, since it does take a value
-/// (`ffplay`'s `-? topic`) with the ordinary space.
-fn spelling_is_sigil(flag: &Entity) -> bool {
-    flag.spellings.len() == 1
-        && matches!(flag.spellings[0].dashes, Dashes::None)
-        && flag.spellings[0]
-            .name
-            .chars()
-            .next()
-            .is_some_and(|c| !c.is_alphanumeric())
-}
-
-/// One entity's spellings, value placeholder, and description — styled
-/// per spec §9.2 (spelling: accent; value: muted; description: default)
-/// — laid out against the section's shared column. Every description
-/// line starts at that column, except a head that reaches it: it keeps
-/// its own line and its first description line starts one space past,
-/// with every continuation back at the column (spec §9.3). Returned lines
-/// are one logical row; the caller records that as an [`EntryRow`].
-fn entity_line(
-    flag: &Entity,
-    dim: bool,
-    width: usize,
-    color_enabled: bool,
-    // Where the value and description columns begin, shared across the
-    // whole section — see `section_layout`.
-    layout: SectionLayout,
-) -> Vec<Line<'static>> {
-    let name_spec = entity_name_spec(flag);
-    let value_text = entity_value_text(flag);
-
-    // Two columns, chosen by the row's own shape (spec §9.3), behind the
-    // section's own indent: shorts at the content edge, longs preindented
-    // so every long in the section starts in the same place whether or not
-    // a short precedes it.
-    let head_column = spelling_column(flag, layout.indent);
-    let leading = " ".repeat(head_column);
-    let leading = leading.as_str();
-    let spelling_style = if dim {
-        style::muted(color_enabled)
-    } else {
-        style::accent(color_enabled)
-    };
-    // Muted, not italic: many terminals silently ignore italic (spec §9.2).
-    let value_style = style::muted(color_enabled);
-    let desc_style = if dim {
-        style::muted(color_enabled)
-    } else {
-        Style::default()
-    };
-
-    let mut first_line_spans = vec![Span::styled(
-        format!("{leading}{name_spec}"),
-        spelling_style,
-    )];
-    let mut prefix_width = display_width(leading) + display_width(&name_spec);
-    if let Some(v) = &value_text {
-        // One space after the spelling, never a padded slot of its own
-        // (spec §9.3), except the argfile sigil flag (spec §4.5, `@<file>`).
-        let gap = if spelling_is_sigil(flag) { "" } else { " " };
-        first_line_spans.push(Span::raw(gap));
-        first_line_spans.push(Span::styled(v.clone(), value_style));
-        prefix_width += display_width(gap) + display_width(v);
-    }
-
-    // A head wider than the pane is broken across lines here rather than
-    // handed to `Paragraph`'s defensive `Wrap` (module doc), which restarts
-    // at column 0 with no memory of the row's indent. Regression fixture:
-    // `vgchange --alloc` rendered through a real pty (AGENTS.md §3.2).
-    let mut head: Vec<Line<'static>> = Vec::new();
-    if prefix_width > width {
-        // Budget the wrap at the row's own column, so every line of the
-        // head sits in the area reserved for heads rather than drifting
-        // into the description's.
-        let budget = width.saturating_sub(head_column).max(1);
-        let indent = leading.to_string();
-        for (i, chunk) in wrap_words(&name_spec, budget).into_iter().enumerate() {
-            let text = if i == 0 {
-                format!("{leading}{chunk}")
-            } else {
-                format!("{indent}{chunk}")
-            };
-            head.push(Line::from(Span::styled(text, spelling_style)));
-        }
-        if let Some(v) = &value_text {
-            for chunk in wrap_words(v, budget) {
-                head.push(Line::from(Span::styled(
-                    format!("{indent}{chunk}"),
-                    value_style,
-                )));
-            }
-        }
-    }
-
-    let deprecated_tag = flag
-        .deprecated
-        .as_ref()
-        .map(|_| " (deprecated)".to_string());
-
-    let mut description_text = flag.description.as_ref().map(|d| d.single_line());
-
-    if let Some(tag) = &deprecated_tag {
-        description_text = Some(match description_text {
-            Some(d) => format!("{d}{tag}"),
-            None => tag.trim_start().to_string(),
-        });
-    }
-    let description_text = description_text.filter(|d| !d.is_empty());
-
-    // A flag's `choices` (spec §7 Tier B rule 4) never fold into
-    // `description_text` or the spelling: they render as their own
-    // `values:` line, indented two past the description column (spec
-    // §9.3). Choices with no per-value description (e.g. `tar
-    // --quoting-style`) get the single summary line below; choices that
-    // carry their own text (ffmpeg/ffplay AVOptions) render one indented
-    // `name  description` row each instead, via `choice_detail_lines`.
-    let has_choice_descriptions = flag.choices.iter().any(|c| c.description.is_some());
-    let values_line = (!flag.choices.is_empty() && !has_choice_descriptions).then(|| {
-        let joined = flag
-            .choices
-            .iter()
-            .map(|c| c.name.as_str())
-            .collect::<Vec<_>>()
-            .join(", ");
-        format!("values: {joined}")
-    });
-
-    if description_text.is_none() && values_line.is_none() && !has_choice_descriptions {
-        if !head.is_empty() {
-            return head;
-        }
-        return vec![Line::from(first_line_spans)];
-    }
-
-    // One description column for the entire section (spec §9.3). A head
-    // that reaches the column keeps its own line, never truncated (spec
-    // §9.1); only its first description line starts one space past it,
-    // every later line is back at the column.
-    let column = layout.description;
-    let rest_width = width.saturating_sub(column).max(1);
-
-    let mut lines = Vec::new();
-    let mut remainder = description_text.clone().unwrap_or_default();
-    if head.is_empty() {
-        if description_text.is_some() {
-            // One space past the head, or the column — whichever is
-            // further right. They coincide for every row whose head fits.
-            let start = column.max(prefix_width + 1);
-            let first = width
-                .checked_sub(start)
-                .and_then(|room| leading_words(&remainder, room));
-            if let Some((first_chunk, rest)) = first {
-                first_line_spans.push(Span::raw(" ".repeat(start - prefix_width)));
-                first_line_spans.push(Span::styled(first_chunk, desc_style));
-                remainder = rest;
-            }
-        }
-        // No description: the head is the whole first line, and a
-        // choices-only flag's values line still lands on the next one.
-        lines.push(Line::from(first_line_spans));
-    } else {
-        // A head broken across lines has taken the pane's whole width for
-        // itself; its description begins on the line after it, at the
-        // column like any other.
-        lines.extend(head);
-    }
-
-    let indent_str = " ".repeat(column);
-    if !remainder.is_empty() {
-        for chunk in wrap_words(&remainder, rest_width) {
-            lines.push(Line::from(Span::styled(
-                format!("{indent_str}{chunk}"),
-                desc_style,
-            )));
-        }
-    }
-
-    // The values line sits two columns past the description column,
-    // marking it subordinate rather than a second row (spec §9.3).
-    if let Some(values_line) = values_line {
-        let values_column = column + 2;
-        let values_width = width.saturating_sub(values_column).max(1);
-        let values_indent = " ".repeat(values_column);
-        let values_style = style::muted(color_enabled);
-        for chunk in wrap_words(&values_line, values_width) {
-            lines.push(Line::from(Span::styled(
-                format!("{values_indent}{chunk}"),
-                values_style,
-            )));
-        }
-    } else if has_choice_descriptions {
-        lines.extend(choice_detail_lines(flag, column, width, color_enabled));
-    }
-
-    lines
-}
-
-/// Render a flag's choices as a `values:` header followed by one indented
-/// `name  description` row per choice (spec §9.3/§4.1's render note),
-/// dim/derived style matching the bare `values:` line this replaces.
-///
-/// Only called when at least one choice carries a description — the mixed
-/// case (some choices described, some bare, e.g. ffmpeg's `-bug`, whose
-/// `autodetect` value has no text of its own) still renders every choice
-/// through this path so the list stays one table rather than splitting
-/// into two.
-fn choice_detail_lines(
-    flag: &Entity,
-    column: usize,
-    width: usize,
-    color_enabled: bool,
-) -> Vec<Line<'static>> {
-    let values_column = column + 2;
-    let values_indent = " ".repeat(values_column);
-    let values_style = style::muted(color_enabled);
-
-    let choice_column = values_column + 2;
-    let choice_indent = " ".repeat(choice_column);
-    let choice_width = width.saturating_sub(choice_column).max(1);
-    let name_width = flag
-        .choices
-        .iter()
-        .map(|c| display_width(&c.name))
-        .max()
-        .unwrap_or(0);
-
-    let mut lines = vec![Line::from(Span::styled(
-        format!("{values_indent}values:"),
-        values_style,
-    ))];
-
-    for choice in &flag.choices {
-        let desc = choice
-            .description
-            .as_ref()
-            .map(|d| d.single_line())
-            .filter(|d| !d.is_empty());
-        let Some(desc) = desc else {
-            lines.push(Line::from(Span::styled(
-                format!("{choice_indent}{}", choice.name),
-                values_style,
-            )));
-            continue;
-        };
-        let head = format!("{:<name_width$}  ", choice.name);
-        let head_width = display_width(&head);
-        let first_room = choice_width.saturating_sub(head_width);
-        match leading_words(&desc, first_room) {
-            Some((first_chunk, rest)) => {
-                lines.push(Line::from(Span::styled(
-                    format!("{choice_indent}{head}{first_chunk}"),
-                    values_style,
-                )));
-                // `leading_words` returns `rest == ""` when the whole
-                // description fit on the first line — `wrap_words` always
-                // returns at least one (possibly empty) chunk for callers
-                // that need one, which is exactly wrong here: an empty
-                // `rest` means there is no continuation to render, not one
-                // blank line's worth. Skipping the call entirely (rather
-                // than filtering its output) is what keeps a choice whose
-                // description merely happens to be blank-after-wrapping
-                // indistinguishable from one that never had a remainder —
-                // there is no such case, `rest` is only ever the literal
-                // empty string when nothing is left.
-                if !rest.is_empty() {
-                    let cont_indent = " ".repeat(choice_column + head_width);
-                    let cont_width = width.saturating_sub(choice_column + head_width).max(1);
-                    for chunk in wrap_words(&rest, cont_width) {
-                        lines.push(Line::from(Span::styled(
-                            format!("{cont_indent}{chunk}"),
-                            values_style,
-                        )));
-                    }
-                }
-            }
-            None => {
-                lines.push(Line::from(Span::styled(
-                    format!("{choice_indent}{}", choice.name),
-                    values_style,
-                )));
-                for chunk in wrap_words(&desc, choice_width) {
-                    lines.push(Line::from(Span::styled(
-                        format!("{choice_indent}{chunk}"),
-                        values_style,
-                    )));
-                }
-            }
-        }
-    }
-
-    lines
-}
-
-/// The words of `text` that fit in `width` display columns, and what is
-/// left of it — `None` when not even the first word fits.
-///
-/// The greedy first line of a description whose head pushed it right
-/// (spec §9.3). A word too wide for the room beside the head is not broken
-/// there: `None` sends the whole description to the next line, where the
-/// section's column gives it the room to break in. Splitting it twice —
-/// once against a few leftover cells, once against the column — is how a
-/// pushed row ends up less readable than a plain one.
-fn leading_words(text: &str, width: usize) -> Option<(String, String)> {
-    let mut first = String::new();
-    let mut used = 0usize;
-    let mut words = text.split_whitespace();
-    let mut rest: Vec<&str> = Vec::new();
-
-    for word in words.by_ref() {
-        let candidate = used + usize::from(used > 0) + display_width(word);
-        if candidate > width {
-            rest.push(word);
-            break;
-        }
-        if used > 0 {
-            first.push(' ');
-        }
-        first.push_str(word);
-        used = candidate;
-    }
-    if first.is_empty() {
-        return None;
-    }
-    rest.extend(words);
-    Some((first, rest.join(" ")))
 }
 
 /// The provenance footer (spec §2, §4.2): which sources contributed, and
